Refactor bulkSplitter
Simplified looping. Also, we no longer truncate mid-line unless it's the first
line of a chunk (in such a case we have no option but to truncate). A truncated
line will be followed by an ellipsis. Otherwise we're splitting chunks at the
last possible newline without exceeding charLimit for that chunk.