Rework #146. Fixes various errors relating to input length. (#150)
* Revert "Fix notify silently fails (#146)"

This reverts commit 173f91472fd73c410969d063776ff1a7224fd6d1.

* Emit logged lines to stdout, not stderr

Fixes #134 (this was previously fixed but got backed out in landing #130)

* Split lines as per CharLimit in non-bulk mode

Resolves #137

* Bubble up any Scanner errors

This surfaces the error that's causing #138:

'bufio.Scanner: token too long'

* Fix infinite loop with -char-limit <= len('...')

We're using '...' to indicate that a line has been truncated. If -char-limit
was less than the length of this ellipsis string, the scanner would never
terminate.

Raise an error if the charLimit given to a splitter is <= len('...')

* Ensure we never allow the Scanner buffer to fill

We know we never need more than CharLimit chars from the Scanner in one go

First of all, we ensure that the Scanner has a buffer which can hold at least
CharLimit chars.

Then we handle cases where the Scanner wants more data but we don't need it
to get more data. Thus it should never end up in a place where its internal
buffer is filled, and it should never return bufio.ErrToLong

Fixes #138

* fmt.print => gologger

Co-authored-by: mzack <marco.rivoli.nvh@gmail.com>